+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2021.06.09
 
--   Add Deposits struct
+-   Add Deposits struct(#206)
 -   Change the value type of AccountCollateral from Balance to Deposits 
 -   Rename AccountCollateral to AccountDeposits
 -   Remove the storage AccountCollateralAsset